# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/terpenes": {
      "get": {
        "description": "Returns the full terpene reference catalog ordered alphabetically by Hebrew name. The list is small (~17 rows) and intended to be cached client-side.",
        "operationId": "TerpeneController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Catalog fetched successfully. `result` is an array of TerpeneDto ordered alphabetically by name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerpeneListResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Not applicable for this endpoint."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all terpenes",
        "summaryHe": "מציג את קטלוג הטרפנים המלא המוגדר במערכת",
        "tags": [
          "terpenes"
        ],
        "toolIcon": "ph-flower-lotus"
      },
      "post": {
        "description": "Creates a new terpene entry in the catalog. The `name` field must be unique — attempting to create a duplicate will return a 409 Conflict. The `color` field must be a valid hex color (e.g., #66BB6A).",
        "operationId": "TerpeneController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Terpene creation payload. `name` and `color` are required; `description`, `scent`, and `effects` are optional.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerpeneCreateDto"
              },
              "examples": {
                "example1": {
                  "summary": "Typical terpene creation",
                  "value": {
                    "name": "לימונן",
                    "description": "טרפן הדרי עם ניחוח לימון",
                    "scent": "הדרים, לימון",
                    "effects": [
                      "מרומם",
                      "ממריץ"
                    ],
                    "color": "#FFEB3B"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Terpene created successfully. `result` is the created terpene.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerpeneResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. Check that `name` is provided, `color` is a valid hex, and `effects` are strings."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "409": {
            "description": "A terpene with the same name already exists."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new terpene",
        "summaryHe": "יוצר טרפן חדש בקטלוג המערכת",
        "tags": [
          "terpenes"
        ],
        "toolIcon": "ph-flower-lotus"
      }
    },
    "/terpenes/{name}": {
      "get": {
        "description": "Returns the terpene whose `name` column exactly matches the supplied `:name` path parameter. Result is `null` when no record matches — this is not treated as a 404 so the frontend can render a graceful empty state.",
        "operationId": "TerpeneController_findOne",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew terpene name. Must exactly match the `terpene.name` column.",
            "schema": {
              "example": "מירצן",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lookup succeeded. `result` is the matching terpene or null when no row matches the given name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerpeneResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Not applicable for this endpoint."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint. Missing rows return `result: null`."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a terpene by name",
        "summaryHe": "מציג פרטים מלאים על טרפן ספציפי לפי שמו",
        "tags": [
          "terpenes"
        ],
        "toolIcon": "ph-flower-lotus"
      },
      "patch": {
        "description": "Updates an existing terpene. Only the provided fields are modified — omitted fields retain their current values. The `name` cannot be changed via this endpoint; use POST /terpenes to create a new record with a different name.",
        "operationId": "TerpeneController_update",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew terpene name. Must exactly match the `terpene.name` column.",
            "schema": {
              "example": "מירצן",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Terpene update payload. All fields are optional; only provided fields will be updated. `color` must be a valid hex color if provided.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerpeneUpdateDto"
              },
              "examples": {
                "example1": {
                  "summary": "Partial update",
                  "value": {
                    "description": "Updated description",
                    "color": "#4CAF50"
                  }
                },
                "example2": {
                  "summary": "Clear optional fields",
                  "value": {
                    "description": null,
                    "scent": null,
                    "effects": []
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terpene updated successfully. `result` is the updated terpene.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerpeneResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. Check that `color` is a valid hex if provided."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "No terpene matches the given name."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a terpene by name",
        "summaryHe": "מעדכן את המאפיינים, הריח וההשפעות של טרפן קיים",
        "tags": [
          "terpenes"
        ],
        "toolIcon": "ph-flower-lotus"
      },
      "delete": {
        "operationId": "TerpeneController_delete",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew terpene name to delete.",
            "schema": {
              "example": "מירצן",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Terpene deleted successfully."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "404": {
            "description": "No terpene matches the given name."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a terpene by name.",
        "summaryHe": "מוחק טרפן לצמיתות מקטלוג המערכת",
        "tags": [
          "terpenes"
        ],
        "toolIcon": "ph-trash"
      }
    },
    "/terpenes/{name}/enrich": {
      "post": {
        "description": "Searches the web for the given terpene name, sends context to LLM, and returns enriched description with web results. Does not persist — caller decides whether to save.",
        "operationId": "TerpeneController_enrichSingle",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew terpene name to enrich.",
            "schema": {
              "example": "מירצן",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enrichment succeeded. `result` contains the LLM-generated description and web search results."
          },
          "400": {
            "description": "Invalid terpene name."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "No terpene matches the given name."
          },
          "500": {
            "description": "LLM or web search failure."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enrich a single terpene",
        "summaryHe": "מעשיר טרפן בודד בפרטי ארומה והשפעות מבוססי AI",
        "tags": [
          "terpenes"
        ],
        "toolIcon": "ph-flower-lotus"
      }
    },
    "/terpenes/enrich-missing": {
      "post": {
        "operationId": "TerpeneController_enrichMissing",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bulk enrichment completed."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "500": {
            "description": "LLM or web search failure."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enrich all terpenes with missing properties (description, scent, effects).",
        "summaryHe": "מפעיל סריקה והעשרה אוטומטית לכל הטרפנים שחסר להם מידע מבוסס AI",
        "tags": [
          "terpenes"
        ],
        "toolIcon": "ph-magic-wand"
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TerpeneDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Auto-incremented primary key.",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Hebrew display name. Unique — used as the lookup key from the frontend.",
            "example": "מירצן"
          },
          "englishName": {
            "type": "string",
            "description": "English name of the terpene. Omitted when not set.",
            "example": "Myrcene"
          },
          "description": {
            "type": "string",
            "description": "Short Hebrew description of the terpene. Omitted when not set.",
            "example": "הטרפן הנפוץ ביותר בקנאביס"
          },
          "scent": {
            "type": "string",
            "description": "Aroma profile of the terpene in Hebrew. Omitted when not set.",
            "example": "אדמה, פירות יער"
          },
          "effects": {
            "description": "Effect tags associated with the terpene (Hebrew short labels).",
            "example": [
              "מרגיע",
              "משכך כאבים"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "color": {
            "type": "string",
            "description": "Hex accent color used by the frontend to render per-terpene UI accents (dots, tag borders, etc.).",
            "example": "#66BB6A"
          },
          "colorDark": {
            "type": "string",
            "description": "WCAG AA-safe variant of `color` for dark theme backgrounds (#080D1A).",
            "example": "#66BB6A"
          },
          "colorLight": {
            "type": "string",
            "description": "WCAG AA-safe variant of `color` for light theme backgrounds (#F0F4F8).",
            "example": "#2E7D32"
          }
        },
        "required": [
          "id",
          "name",
          "color",
          "colorDark",
          "colorLight"
        ]
      },
      "TerpeneListResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable status message.",
            "example": "OK"
          },
          "result": {
            "description": "The list of terpenes ordered alphabetically by name.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TerpeneDto"
            }
          },
          "error": {
            "description": "Optional error details.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "retryAfter": {
            "type": "string",
            "description": "Optional retry hint."
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code echoed from the server.",
            "example": 200
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "TerpeneResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable status message.",
            "example": "OK"
          },
          "result": {
            "description": "The matching terpene, or null when no record matches the given name.",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TerpeneDto"
              }
            ]
          },
          "error": {
            "description": "Optional error details.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "retryAfter": {
            "type": "string",
            "description": "Optional retry hint."
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code echoed from the server.",
            "example": 200
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "TerpeneCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Hebrew display name. Must be unique. Used as the lookup key from the frontend.",
            "example": "מירצן"
          },
          "englishName": {
            "type": "string",
            "description": "English name of the terpene.",
            "example": "Myrcene"
          },
          "description": {
            "type": "string",
            "description": "Short Hebrew description of the terpene.",
            "example": "הטרפן הנפוץ ביותר בקנאביס"
          },
          "scent": {
            "type": "string",
            "description": "Aroma profile of the terpene in Hebrew.",
            "example": "אדמה, פירות יער"
          },
          "effects": {
            "description": "Effect tags associated with the terpene (Hebrew short labels).",
            "example": [
              "מרגיע",
              "משכך כאבים"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "color": {
            "type": "string",
            "description": "Hex accent color used by the frontend to render per-terpene UI accents (dots, tag borders, etc.). Must be a valid hex color starting with #.",
            "example": "#66BB6A"
          }
        },
        "required": [
          "name",
          "color"
        ]
      },
      "TerpeneUpdateDto": {
        "type": "object",
        "properties": {
          "englishName": {
            "type": "string",
            "description": "English name of the terpene. Pass null or empty string to clear.",
            "example": "Myrcene"
          },
          "description": {
            "type": "string",
            "description": "Short Hebrew description of the terpene. Pass null or empty string to clear.",
            "example": "הטרפן הנפוץ ביותר בקנאביס"
          },
          "scent": {
            "type": "string",
            "description": "Aroma profile of the terpene in Hebrew. Pass null or empty string to clear.",
            "example": "אדמה, פירות יער"
          },
          "effects": {
            "description": "Effect tags associated with the terpene (Hebrew short labels). Pass empty array to clear.",
            "example": [
              "מרגיע",
              "משכך כאבים"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "color": {
            "type": "string",
            "description": "Hex accent color used by the frontend to render per-terpene UI accents (dots, tag borders, etc.). Must be a valid hex color starting with #.",
            "example": "#66BB6A"
          }
        }
      },
...
    }
  }
}